fix: no completions when deleting text
When diffing text, sometimes this situation may occur:
OLD:
    fields.Many2one('fubuzilla')<EOF>
                        cursor^
NEW:
    fields.Many2one('')<EOF>
               cursor^
DELTA:
    fields.Many2one('fubuzilla')<EOF>
               start^      end^

In this situation, because the rope was updated first, the delta that
the client sends will have its range-end fall out of bounds of the
updated rope. This commit passes in the old rope when diffing text
so that the range-end can always be parsed correctly.