fix: edge case with model fields
When a comment is the first element in an argument list, like this:

    foo = fields.One2many( # haha funny comment
        'res.partner',
        .. # rest of the field
    )

The LSP won't be able to detect the relation at all. Since this edge
case affects an important field, this commit fixes it but this is very
much an ugly hack. There is no alternative solution other than handwriting
a looping parser subroutine.